updated runtime type to float

+++ src/simulators/result.py
@@ -1,6 +1,6 @@
 class RuntimeResult:
     """Defines a list of simulation steps, runtimes, and actions"""
-    def __init__(self, runtime: int=0, label: str=None):
+    def __init__(self, runtime: float=0, label: str=None):
         self.history = list() if runtime == 0 and label is None else [(runtime, label)]
         self.runtime = runtime
